refactor: Renaming folders
