+++ py-scripts/toolbelt/v2/command.py
@@ -4,7 +4,7 @@ from toolbelt.constants import INTERNAL_DIR
 from toolbelt.v2.artifacts.launcher import release_launcher
 from toolbelt.v2.artifacts.player import copy_players
 from toolbelt.v2.github.parser import latest_rc_tags
-from toolbelt.v2.manifests.update import update_manifests
+from toolbelt.k8s_update.update import update_manifests
 from toolbelt.v2.planet.apv import generate_internal_apv, get_old_internal_apv
 from toolbelt.v2.types import Mode
 
@@ -17,7 +17,9 @@ def internal_release(tag: str, mode: Mode = ""):
     # get tag, sha from repository
     launcher_tag, launcher_sha = latest_rc_tags("9c-launcher", mode=mode)
     player_tag, player_sha = latest_rc_tags("NineChronicles", mode=mode)
-    headless_tag, headless_sha = latest_rc_tags("NineChronicles.Headless", mode=mode)
+    headless_tag, headless_sha = latest_rc_tags(
+        "NineChronicles.Headless", mode=mode
+    )
 
     print("Tags", launcher_tag, player_tag, headless_tag)
     print()
